backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smart-nutrition-platform"
version = "0.1.0"
description = "智能营养建议平台后端服务"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "pyjwt>=2.8.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "boto3>=1.34.0",
    "openai>=1.0.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "aiosmtplib>=3.0.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.92.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "alembic/*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
